+++ Changelog.md
@@ -1,5 +1,13 @@
 # `hmdq` Change Log
 
+## 0.3.0
+*August 4th, 2019*
+
+* Added an option for anonymization of sensitive info (basically serial numbers of the tracked devices).
+* Added new config values to support anonymization.
+
+>Note: Upgrading users should regenerate the default config and then merge the old one.
+
 ## 0.2.4
 *August 2nd, 2019*